1.1.3: No minimum weight
Having a minimum weight pushes new lessons before reviews
what might not be what some users want.

If a user wants to make new lessons, it's easy to choose to
do so manually, but finding a lesson to review when the script
asks you do do a new one, is not as easy.